chore(sdk): minor clean up of skipped tests (#7154)
