+++ tests/pytest_tests/system_tests/test_artifacts/test_wandb_run.py
@@ -1,4 +1,5 @@
 import json
+import os
 from unittest import mock
 
 import pytest
@@ -354,3 +355,21 @@ def test_log_code_env(
             else:
                 with pytest.raises(wandb.errors.CommError):
                     wandb.Api().artifact(f"{artifact_name}:v0")
+
+
+@pytest.mark.xfail(reason="Backend race condition")
+def test_anonymous_mode_artifact(wandb_init, capsys, local_settings):
+    copied_env = os.environ.copy()
+    copied_env.pop("WANDB_API_KEY")
+    copied_env.pop("WANDB_USERNAME")
+    copied_env.pop("WANDB_ENTITY")
+    with mock.patch.dict("os.environ", copied_env, clear=True):
+        run = wandb_init(anonymous="must")
+        run.log_artifact(wandb.Artifact("my-arti", type="dataset"))
+        run.finish()
+
+    _, err = capsys.readouterr()
+
+    assert (
+        "Artifacts logged anonymously cannot be claimed and expire after 7 days." in err
+    )
