+++ tests/pytest_tests/system_tests/test_core/test_wandb.py
@@ -173,24 +173,6 @@ def test_anonymous_mode(wandb_init, capsys, local_settings):
     )
 
 
-@pytest.mark.xfail(reason="Backend race condition")
-def test_anonymous_mode_artifact(wandb_init, capsys, local_settings):
-    copied_env = os.environ.copy()
-    copied_env.pop("WANDB_API_KEY")
-    copied_env.pop("WANDB_USERNAME")
-    copied_env.pop("WANDB_ENTITY")
-    with mock.patch.dict("os.environ", copied_env, clear=True):
-        run = wandb_init(anonymous="must")
-        run.log_artifact(wandb.Artifact("my-arti", type="dataset"))
-        run.finish()
-
-    _, err = capsys.readouterr()
-
-    assert (
-        "Artifacts logged anonymously cannot be claimed and expire after 7 days." in err
-    )
-
-
 def test_run_id(wandb_init):
     with mock.patch.dict("os.environ", {"WANDB_RUN_ID": "123456"}):
         run = wandb_init()
